+++ backend/blockchain/test/Manual_Entry_Test.sol
@@ -6,21 +6,18 @@ import "truffle/DeployedAddresses.sol";
 import "../contracts/Manual_Entry.sol";
 import "../contracts/Manual_Entry_Tracker.sol";
 
-contract Manual_Entry_Test
-{
-    // The address of the Module contract to be tested
+contract Manual_Entry_Test {
     Manual_Entry man_entry = Manual_Entry(DeployedAddresses.Manual_Entry());
 
     Manual_Entry_Tracker man_entry_tracker = Manual_Entry_Tracker(DeployedAddresses.Manual_Entry_Tracker());
 
-    // The expected length of the array 
-    uint expectedSecondLength = 2;
+    uint256 expectedSecondLength = 2;
 
-    uint expectedFirstLength = 1;
+    uint256 expectedFirstLength = 1;
 
-    uint expectedSecondIndex = 1;
+    uint256 expectedSecondIndex = 1;
 
-    uint expectedFirstIndex = 0;
+    uint256 expectedFirstIndex = 0;
 
     string expectedSecondKey = "This is the expected key";
 
@@ -30,87 +27,90 @@ contract Manual_Entry_Test
 
     string expectedFirstString = "This is a sample string";
 
-    function testAddHash() public 
-    {
+    //Test whether a hash is being added correctly:
+    //  -- The size of the tracker should be correct 
+    //  -- The index being added to the tracker should be correct
+    function testAddHash() public {
         man_entry.addHash(expectedFirstString);
 
-        uint returnedIndex = man_entry.getLastIndex();
+        uint256 returnedIndex = man_entry.getLastIndex();
 
         man_entry_tracker.addTracker(expectedFirstKey, returnedIndex);
 
-        uint size = man_entry_tracker.getLength();
+        uint256 size = man_entry_tracker.getLength();
 
         Assert.equal(size, expectedFirstLength, "Expected size should match the size that is returned.");
 
         Assert.equal(returnedIndex, expectedFirstIndex, "Expected Index should match the index that is returned.");
     }
-
-    function testAddSecondHash() public
-    {
+    
+    //Test whether a second hash is being added correctly:
+    //  -- The size of the tracker should be correct 
+    //  -- The index being added to the tracker should be correct
+    function testAddSecondHash() public {
         man_entry.addHash(expectedSecondString);
 
-        uint returnedIndex = man_entry.getLastIndex();
+        uint256 returnedIndex = man_entry.getLastIndex();
 
         man_entry_tracker.addTracker(expectedSecondKey, returnedIndex);
 
-        uint size = man_entry_tracker.getLength();
+        uint256 size = man_entry_tracker.getLength();
 
         Assert.equal(size, expectedSecondLength, "Expected size should match the size that is returned.");
 
         Assert.equal(returnedIndex, expectedSecondIndex, "Expected Index should match the index that is returned.");
     }
 
-    function  testLength() public
-    {
-        uint returnedLength = man_entry.getLength();
+    //Test whether the length of the hash array is correct
+    function testLength() public {
+        uint256 returnedLength = man_entry.getLength();
 
         Assert.equal(returnedLength, expectedSecondLength, "Expected length should match the length that is returned.");
     }
 
-    function testReturnFirstHash() public
-    {
-        uint returnedIndex = man_entry_tracker.returnIndex(expectedFirstKey);
+    //Test whether the first hash being returned is correct
+    function testReturnFirstHash() public {
+        uint256 returnedIndex = man_entry_tracker.returnIndex(expectedFirstKey);
 
         string memory returnedHash = man_entry.returnHash(returnedIndex);
-        
+
         Assert.equal(returnedHash, expectedFirstString, "Expected string should match the string that is returned.");
     }
 
-    function testReturnSecondHash() public
-    {
-        uint returnedIndex = man_entry_tracker.returnIndex(expectedSecondKey);
+    //Test whether the second hash being returned is correct
+    function testReturnSecondHash() public {
+        uint256 returnedIndex =
+            man_entry_tracker.returnIndex(expectedSecondKey);
 
         string memory returnedHash = man_entry.returnHash(returnedIndex);
-        
+
         Assert.equal(returnedHash, expectedSecondString, "Expected string should match the string that is returned.");
     }
 
-    function testExists() public
-    {
+    //Test whether the a particular key exists
+    function testExists() public {
         bool exists = man_entry_tracker.checkExists(expectedFirstKey);
 
         Assert.equal(exists, true, "Expected state of existence should match the state of existence that is returned.");
     }
 
-    function testNotExist() public
-    {
+    //Test whether a particular key does not exist
+    function testNotExist() public {
         bool exists = man_entry_tracker.checkExists("does not exist");
 
         Assert.equal(exists, false, "Expected state of existence should match the state of existence that is returned.");
     }
 
-    function testUpdateValue() public
-    {
+    //Test whether the tracked index for a key is changeable, and whether the returned hash is then correct.
+    function testUpdateValue() public {
         man_entry_tracker.addTracker(expectedSecondKey, expectedFirstIndex);
 
-        uint returnedIndex = man_entry_tracker.returnIndex(expectedSecondKey);
+        uint256 returnedIndex = man_entry_tracker.returnIndex(expectedSecondKey);
 
         Assert.equal(returnedIndex, expectedFirstIndex, "Expected index should match the index that is returned.");
 
         string memory returnedHash = man_entry.returnHash(returnedIndex);
-        
+
         Assert.equal(returnedHash, expectedFirstString, "Expected string should match the string that is returned.");
     }
 }
-
-
